// src/app/common/models/export/data-type.enum.ts
export enum DataType {
  /// <summary>
        /// mặc định
        /// </summary>
        DefaultType = 0,

        /// <summary>
        /// Số lượng
        /// </summary>
        QuantityType = 1,

        /// <summary>
        /// đơn giá
        /// </summary>
        UnitPriceType = 2,

        /// <summary>
        /// thành tiền
        /// </summary>
        CurrencyType = 3,

        /// <summary>
        /// kiểu ngày
        /// </summary>
        DateType = 4,

        /// <summary>
        /// kiểu ngày giờ
        /// </summary>
        DateTimeType = 5,

        /// <summary>
        /// kiểu ngày giờ
        /// </summary>
        DateTimeDetailType = 16,

        /// <summary>
        /// Năm
        /// </summary>
        YearType = 6,

        /// <summary>
        /// Kiểu giờ
        /// </summary>
        TimeType = 7,

        /// <summary>
        /// kiểu số
        /// </summary>
        NumberType = 8,

        /// <summary>
        /// Kiểu bool checkbox
        /// </summary>
        CheckBoxType = 9,

        /// <summary>
        /// Kiểu hyperlink
        /// </summary>
        HyperLinkType = 10,

        /// <summary>
        /// phần trăm
        /// </summary>
        PercentType = 11,

        /// <summary>
        /// kiểu liệt kê
        /// </summary>
        EnumType = 12,

        /// <summary>
        /// Hệ số tỉ lệ
        /// </summary>
        ExchangeRateType = 13,

        /// <summary>
        /// sắp xếp
        /// </summary>
        SortOrderType = 14,

        /// <summary>
        /// Không định dạng
        /// </summary>
        NoFormat = 15,
        /// <summary>
        /// Không định dạng
        /// </summary>
        DecimalType = 18
}